backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
adversarial-robustness-toolbox==1.13.1
torch
torchvision
numpy
matplotlib
Pillow
requests
drf-yasg
pycryptodome
